- When getting the cursor set it to start reading from the requested token   - Had to backport the new EventSequenceToken (called it SimpleSequenceToken) for this to work because SequenceNumber property needs to be public. - Delete from Redis when doing a pop operation in RedisCustomList

+++ CustomPersistentStreamProvider/RedisCache/ICacheList.cs
@@ -5,13 +5,13 @@ namespace PipeStreamProvider.RedisCache
 {
     public interface ICacheList<T>
     {
-        T Get(int index);
+        T Get(long index);
         T LeftPop();
         bool LeftPush(T v);
         T RightPop();
         bool RightPush(T v);
-        bool Set(int index, T newVal);
+        bool Set(long index, T newVal);
 
-        int Count { get; }
+        long Count { get; }
     }
 }
